% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcap.result.R
\name{isEditMode}
\alias{isEditMode}
\title{Is the page that called this function edit or mini (or derviatives)}
\usage{
isEditMode()
}
\value{
Logical: TRUE if it is the edit page
}
\description{
Retrieves the url from the .session info and regex's it for edit.html
}

