% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controllerR6.R
\name{parseSizesJson}
\alias{parseSizesJson}
\title{Get plot sizes from JSON}
\usage{
parseSizesJson(json)
}
\arguments{
\item{json}{A JSON string from the client. Must have an element named plotSizes}
}
\value{
List with control IDs as names and named vector of widths and heights
as list items.
}
\description{
Parse the JSON string sent from the client, find the plot IDs and sizes, and
return as a neat R object.
}

