% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rcloud.rcap.caps}
\alias{rcloud.rcap.caps}
\title{List of OCAPS}
\format{An object of class \code{NULL} of length 0.}
\usage{
rcloud.rcap.caps
}
\description{
\itemize{
  \item \code{getRFunctions} Query all R functions in the
    rcloud environment. It calls
    \code{\link{rcloud.rcap.global.functions}}.
  \item \code{getRTime} Query the current time of the R interpreter.
    This is used for debugging.
  \item \code{updateControls} Send an update request from the
    front-end to the back-end. It calls the \code{update} method
    of the Controller.
}
}
\keyword{datasets}

